+++ app/src/main/java/com/bongtu/baekseo/data/dto/auth/PostKakaoLoginRequest.kt
@@ -1,4 +1,4 @@
-package com.bongtu.baekseo.data.dto.oauth
+package com.bongtu.baekseo.data.dto.auth
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
